tailwind extension package.json update
